new features
I added several other new features to the pages, such as: a structure for the services page, a new error page in case the user tries to access a page by typing incorrectly, thus redirecting to a home page. There was also styling in everything that was added and metadata for a future search on Google.